Changed the arguments from "int" to "unsigned short".
No factorial would be able to be generated of a full int, partial or not, and
it cannot be negative. To safegaurd against that I made them unsigned and short
so that it will not be able to run outside of those simple ranges.

+++ include/factorial.h
@@ -10,6 +10,6 @@
 
 //  Limited to a product no greater than 18446744073709551615
 unsigned long long
-factorial ( int start , int end ) ;
+factorial ( unsigned short start , unsigned short end ) ;
 
 #endif
